--- a/.github/workflows/build-windows.yml
+++ b/.github/workflows/build-windows.yml
@@ -18,15 +18,14 @@ jobs:
         # Github Actions requires a single row to be added to the build matrix.
         # See https://help.github.com/en/articles/workflow-syntax-for-github-actions.
         name: [
-          #TODO This build keeps timing out, need to understand why.
-          # windows-2016-cl,
-          windows-2019-cl,
-        ]
+            #TODO This build keeps timing out, need to understand why.
+            # windows-2016-cl,
+            windows-2019-cl,
+          ]
 
         build_type: [Debug, Release]
         build_unstable: [ON]
         include:
-
           #TODO This build keeps timing out, need to understand why.
           # - name: windows-2016-cl
           #   os: windows-2016
@@ -37,10 +36,7 @@ jobs:
             compiler: cl
 
     steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Install (Windows)
-        if: runner.os == 'Windows'
+      - name: Install
         shell: powershell
         run: |
           Invoke-Expression (New-Object System.Net.WebClient).DownloadString('https://get.scoop.sh')
@@ -72,8 +68,10 @@ jobs:
           boost_version: 1.72.0
           toolset: msvc14.2
 
-      - name: Build (Windows)
-        if: runner.os == 'Windows'
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Build
         env:
           BOOST_ROOT: ${{ steps.install-boost.outputs.BOOST_ROOT }}
         run: |