Specialize Boost install so that the CI is faster.

This gives a ~17 minute improvement since Boost install goes down from ~22 mins to ~5 mins.

Shoutout to @ProfFan for making me aware of this!