@varunagrawal 2021-08-19
For Linux, not installing all of Boost and using `make -j$(nproc)` gave the biggest savings. For Mac, I did the same trick for `make` but with the appropriate command to figure out number of cores on the system.